AK2: repair unpackelf support on repeat dirty flashes - on Sony devices supporting standard boot.img on unloacked bootloaders first AK2 flash would convert from ELF to boot.img, but second flash would attempt to use unpackelf again and fail, so fallback to unpackbootimg - requires updated unpackelf binary for Sony ELF devices (see: https://github.com/osm0sis/unpackelf)

+++ tools/ak2-core.sh
@@ -32,7 +32,7 @@ dump_boot() {
   else
     dd if=$block of=/tmp/anykernel/boot.img;
   fi;
-  if [ -f "$bin/unpackelf" ]; then
+  if [ -f "$bin/unpackelf" -a "$($bin/unpackelf -i /tmp/anykernel/boot.img -h -q; echo $?)" != 0 ]; then
     $bin/unpackelf -i /tmp/anykernel/boot.img -o $split_img;
     mv -f $split_img/boot.img-ramdisk.cpio.gz $split_img/boot.img-ramdisk.gz;
   else
